requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
av==14.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
coloredlogs==15.0.1
ctranslate2==4.6.0
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.3.0
faster-whisper==1.1.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.2
fsspec==2025.5.1
future==1.0.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-google-genai==2.0.8
langchain-text-splitters==0.3.11
langsmith==0.4.31
google-generativeai==0.8.3
markupsafe==3.0.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
numpy==1.26.2
onnxruntime==1.22.0
openai==1.109.1
opencv-python==4.7.0.72
opencv-python-headless==4.9.0.80
orjson==3.11.3
packaging==25.0
pillow==11.2.1
pip==25.2
proglog==0.1.12
propcache==0.3.1
protobuf==6.31.1
pydantic==2.11.5
pydantic-core==2.33.2
pydub==0.25.1
python-dotenv==1.0.1
pytubefix==9.1.1
pyyaml==6.0.2
regex==2025.9.18
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==65.5.0
sniffio==1.3.1
sqlalchemy==2.0.43
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.15.2
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
webrtcvad-wheels==2.0.14
yarl==1.20.0
zstandard==0.25.0
